* frtypes.h file removed, for any missing types <boost/cstdint.hpp> can be included (now that there's a dependency on Boost)
svn path=/trunk/flamerobin/; revision=2072

+++ src/metadata/database.cpp
@@ -217,7 +217,7 @@ bool DatabaseInfo::getReadOnly() const
 //-----------------------------------------------------------------------------
 int64_t DatabaseInfo::getSizeInBytes() const
 {
-    return (int64_t)getPages() * (int64_t)getPageSize();
+    return static_cast<int64_t>(getPages()) * getPageSize();
 }
 //-----------------------------------------------------------------------------
 int DatabaseInfo::getSweep() const
